Make `NotificationUtils` a class from a file type Also rename it to `NotificationHandler`
